Moves QiskitElementPage component to elements folder
